FIX: final control v1.0

+++ Control_of_Robot_v1/motor_control.c
@@ -91,7 +91,7 @@ void adjustPWM(){
     // pid[2] -> RUEDA 3  
     // pid[3] -> RUEDA 4 
     for(int i = 0 ; i<4; i++){
-        duty[i] =  duty[i] + (int)pid[i];
+        duty[i] =  duty[i] + pid[i];
         if (duty[i] > MAX_DUTY)
         {
             duty[i] = MAX_DUTY;
@@ -118,7 +118,7 @@ void adjustPWM(){
     {
         duty_aux[3] = duty[3];
     }*/
-    setDutyxPID(duty[0], duty[1], duty[2], duty[3]);
+    setDutyxPID((int)duty[0], (int)duty[1], (int)duty[2], (int)duty[3]);
     //printf("%d,%d,%d,%d\n",duty[0], duty[1], duty[2], duty[3]);
 
 }
